+++ DNN Platform/Library/Framework/ServicesFrameworkImpl.cs
@@ -65,19 +65,8 @@ namespace DotNetNuke.Framework
 
         public void RegisterAjaxScript(Page page)
         {
-            var portalSettings = PortalController.Instance.GetCurrentPortalSettings();
-            if (portalSettings == null) return;
-            var path = portalSettings.PortalAlias.HTTPAlias;
-            int index = path.IndexOf('/');
-            if (index > 0)
-            {
-                path = path.Substring(index);
-            }
-            else
-            {
-                path = "/";
-            }
-            path = path.EndsWith("/") ? path : path + "/";
+            var path = ServicesFramework.GetServiceFrameworkRoot();
+
             JavaScript.RegisterClientReference(page, ClientAPI.ClientNamespaceReferences.dnn);
             ClientAPI.RegisterClientVariable(page, "sf_siteRoot", path, /*overwrite*/ true);
             ClientAPI.RegisterClientVariable(page, "sf_tabId", PortalSettings.Current.ActiveTab.TabID.ToString(CultureInfo.InvariantCulture), /*overwrite*/ true);
